requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl_cffi==0.12.0
Flask==3.1.1
Flask-WTF==1.2.2
frozendict==2.4.6
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
multitasking==0.0.12
narwhals==2.0.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
peewee==3.18.2
platformdirs==4.3.8
plotly==6.2.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
soupsieve==2.7
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
yfinance==0.2.65
